requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-multipart==0.0.9
PyMuPDF==1.24.11
trafilatura==1.7.0
fpdf==1.7.2
langchain-groq==1.1.0
python-dotenv==1.2.1
